Address management resolved

+++ Controllers/UserController/addressController.js
@@ -2,7 +2,9 @@ import client from "../../Utils/database.js";
 
 let getAddress = async(req,res)=>{
     try{
-        let result = await client.query('select * from address');
+        let {userId}= req.user;
+        let result = await client.query('select * from address where user_id=$1', [userId]);
+        if(!result) res.send(200).json({});
         let address = result.rows;
         res.status(200).json({address : address});
     }catch(err){
@@ -25,9 +27,10 @@ let postAddress = async (req,res)=>{
 let updateAddress = async(req,res)=>{
     try{
         const id = req.params.id;
+        let {userId} = req.user;
         const {locality, city, state, landmark, alternate_phno, address_type, isDefault} = req.body;
         // let id = await client.query('select id from address where user_id =$1', [userId]);
-        await client.query('update address set locality =$1, city =$2, state =$3, landmark =$4, alternate_phno =$5, address_type =$6, isDefault =$7 where id=$8',[locality, city, state, landmark, alternate_phno, address_type, isDefault, id]);
+        await client.query('update address set locality =$1, city =$2, state =$3, landmark =$4, alternate_phno =$5, address_type =$6, isDefault =$7 where id=$8 and user_id=$9',[locality, city, state, landmark, alternate_phno, address_type, isDefault, id, userId]);
         res.status(200).send('Address Updated Successfully');
     }catch(err){
         res.status(500).send({error:err.message});
@@ -36,10 +39,9 @@ let updateAddress = async(req,res)=>{
 
 let deleteAddress = async(req,res)=>{
     try{
-        // let {userId} = req.user;
-        // let id = await client.query('select id from address where user_id =$1', [userId]);
         const id = req.params.id;
-        await client.query('delete from address where id=$1',[id]);
+        let {userId} = req.user;
+        await client.query('delete from address where id=$1 and user_id =$2',[id,userId]);
         res.status(200).send('Address Deleted Successfully');
     }catch(err){
         res.status(500).send({error:err.message});
